ui: refactor files
